Add Bash highlighting support and refactor highlighting functionality
A separate highlighting logic was introduced in 'hljs.ts' for improved structure and readability. Adding support for bash language became now a straightforward task as illustrated by the added bash import in 'language.ts' and its usage in '+page.svelte'. This change simplifies future additions of new languages and improves