Add visual indicators for removed lines in Highlighted.svelte
This update expands the Highlighted.svelte component to visually indicate lines that have been removed. It adds a new class "removed-line" for this purpose. Furthermore, the omission icon size has been modified, and the span element conditionally added when a line is removed after the current line number. This will make it easier for users to observe the modifications made.